fixed decimal price and liters issue in refuel.

+++ database/migrations/2023_04_25_164511_create_refuels_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('refuels', function (Blueprint $table) {
             $table->id();
-            $table->integer('liters');
+            $table->decimal('liters');
             $table->decimal('cost');
             $table->foreignId('user_id');
             $table->foreignId('car_id');
